created a data folder, added the vehicleList array from game.js in a new file: vehicleList.js, stored in data folded
