+++ astra/src/data/vehicleList.js
@@ -1,0 +1,20 @@
+const vehicleList = [
+  {
+    vehicle: "Car",
+    speed: "120"
+  },
+  {
+    vehicle: "Rocket",
+    speed: "39895"
+  },
+  {
+    vehicle: "Bicyle",
+    speed: "20"
+  },
+  {
+    vehicle: "Camel",
+    speed: "65"
+  }
+];
+
+export default vehicleList
